requirements.txt
# Estructuras químicas y procesamiento de datos
rdkit==2023.3.1  
chembl_structure_pipeline==1.0.7  

# Manipulación y visualización de datos
pandas==2.0.3
numpy==1.24.3  

# Visualización de gráficos y molecular drawing
Pillow==9.5.0  
matplotlib==3.7.1  

# Procesamiento de solicitudes HTTP (descarga de datos)
requests==2.31.0

# Notebooks y visualización en Jupyter
ipython==8.12.0
jupyterlab==4.0.2  

# Biopython para trabajar con archivos MMCIF
biopython==1.81
